direct to error404 for eventedit page

+++ events/eventedit.php
@@ -23,24 +23,48 @@
             $stmt1->bindParam(":TName", $_SESSION['TName']);
             $stmt1->execute();
             $result1=$stmt1->fetch();
+
+            if ($result1['TName']!=$_SESSION['TName']){
+                header("Location: ../error.html");
+            }
             unset($_SESSION['TName']);
             header("Location: team_details.php");
         }
         else if  ($action=="Sdelete"){
-            $query1='DELETE FROM solo_data WHERE P_id=:P_id';
-            $stmt1=$pdo->prepare($query1);
-            $stmt1->bindParam(":P_id", $P_id);
-            $stmt1->execute();
-            $result1=$stmt1->fetch();
+            if ($_SESSION['is_team']==1){
+                $query1='DELETE FROM solo_data WHERE P_id=:P_id and H_id=:H_id and Tuser_id=:user_id ';
+                $stmt1=$pdo->prepare($query1);
+                $stmt1->bindParam(":P_id", $P_id);
+                $stmt1->bindParam(":user_id",$_SESSION['user_id']);
+                $stmt1->bindParam(":H_id",$_SESSION['H_id']);   
+                $stmt1->execute();
+                $result1=$stmt1->fetch();
 
-            $query2="UPDATE team_data SET TMembers=TMembers-1 WHERE TName=:TName and H_id=:H_id and Tuser_id=:user_id ";
-            $stmt2= $pdo->prepare($query2);
-            $stmt2->bindParam(":TName", $_SESSION['TName']);
-            $stmt2->bindParam(":user_id",$_SESSION['user_id']);
-            $stmt2->bindParam(":H_id",$_SESSION['H_id']);   
-            $stmt2->execute();
-            unset($_SESSION['TName']);
-            header("Location: team_details.php");
+                if ($result1['P_id']!= $P_id){
+                    header("Location: ../error.html");
+                }
+                $query2="UPDATE team_data SET TMembers=TMembers-1 WHERE TName=:TName and H_id=:H_id and Tuser_id=:user_id ";
+                $stmt2= $pdo->prepare($query2);
+                $stmt2->bindParam(":TName", $_SESSION['TName']);
+                $stmt2->bindParam(":user_id",$_SESSION['user_id']);
+                $stmt2->bindParam(":H_id",$_SESSION['H_id']);   
+                $stmt2->execute();
+                unset($_SESSION['TName']);
+                header("Location: team_details.php");
+            }
+            else{ 
+                $query1='DELETE FROM solo_data WHERE P_id=:P_id and H_id=:H_id and Puser_id=:user_id ';
+                $stmt1=$pdo->prepare($query1);
+                $stmt1->bindParam(":P_id", $P_id);
+                $stmt1->bindParam(":user_id",$_SESSION['user_id']);
+                $stmt1->bindParam(":H_id",$_SESSION['H_id']);   
+                $stmt1->execute();
+                $result1=$stmt1->fetch();
+                if ($result1['P_id']!= $P_id){
+                    header("Location: ../error.html");
+                }
+                header("Location: member_details.php");
+            }
         }
         else if  ($action=="edit"){
             header("Location: member_details.php");
